+++ app/controllers/v1/observation_documents_controller.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module V1
+  class ObservationDocumentsController < ApiController
+
+    skip_before_action :authenticate, only: [:index, :show]
+    load_and_authorize_resource class: 'Document'
+
+  end
+end